requirements.txt
flask==2.3.3
requests==2.31.0
pillow==10.0.1
pytesseract==0.3.10
mcp-tools==1.0.0
PyPDF2==3.0.1